% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_us_by_cnty.R
\name{cdc_vac_us_by_cnty}
\alias{cdc_vac_us_by_cnty}
\title{Vaccinations for COVID-19 Counts for US Counties from Center for Disease Control (CDC)}
\usage{
cdc_vac_us_by_cnty()
}
\value{
A dataframe/tibble
}
\description{
This functions pulls vaccine data for US counties from the CDC.
  No data is stored within the package, so the data should continuously
  update over time as long as the links do not change.
}
\details{
Website:

    - CDC COVID Tracker: https://covid.cdc.gov/covid-data-tracker/#datatracker-home

    - Data: https://data.cdc.gov/api/views/8xkx-amqh/rows.csv?accessType=DOWNLOAD
}
\examples{
vac_by_cnty <- R.COVID.19::cdc_vac_us_by_cnty()
head(vac_by_cnty)

}
