% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_us_by_state.R
\name{cdc_vac_us_by_state}
\alias{cdc_vac_us_by_state}
\title{Vaccinations for COVID-19 Counts for US States from Center for Disease Control (CDC)}
\usage{
cdc_vac_us_by_state()
}
\value{
A dataframe/tibble
}
\description{
This functions pulls vaccine data for US states from the CDC.
  No data is stored within the package, so the data should continuously
  update over time as long as the links do not change.
}
\details{
Website:

    - CDC COVID Tracker: https://covid.cdc.gov/covid-data-tracker/#datatracker-home

    - CDC State Site: https://data.cdc.gov/Vaccinations/COVID-19-Vaccination-Trends-in-the-United-States-N/rh2h-3yt2

    - Data: https://data.cdc.gov/api/views/rh2h-3yt2/rows.csv?accessType=DOWNLOAD
}
\examples{
vac_by_states <- R.COVID.19::cdc_vac_us_by_state()
head(vac_by_states)

}
