% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_deaths_daily.R
\name{country_deaths_daily}
\alias{country_deaths_daily}
\title{COVID-19 Country Death Counts}
\usage{
country_deaths_daily()
}
\value{
A datarame/tibble
}
\description{
This functions pulls death COVID-19 data from John Hopkins
  GitHub repository downloads it and transpose it so that it is tidy.  No
  data is stored within the package, so the data should continuously update
  over time as long as the links do not change.
}
\examples{
\dontrun{
confirmed <- R.COVID.19::country_deaths_daily()
}

}
