% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_us_co.R, R/covid19_us_st.R
\name{covid19_us_co}
\alias{covid19_us_co}
\title{COVID19 Confirmed Counts US counties}
\usage{
covid19_us_co()

covid19_us_co()
}
\description{
confirmed counts

confirmed counts
}
\details{
grabs the wide makes it long

grabs the wide makes it long
}
