% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covtrck_states_testing.R
\name{covtrck_states_testing}
\alias{covtrck_states_testing}
\title{COVID-19 US Testing Counts from CovidTracking.com}
\usage{
covtrck_states_testing()
}
\value{
A tsibble/dataframe/tibble
}
\description{
This functions pulls US testing counts of COVID-19 data from covidtracking.com.
  The data is tidyed so that it is more user friendly.  No data is stored
  within the package, so the data should continuously update over time as
  long as the links do not change.
}
\details{
Website: https://covidtracking.com/api

  Deprecated fields
  The following fields are deprecated and are no longer being updated with new data.

  /api/v1/states/current.json:

  positiveScore
  negativeScore
  negativeRegularScore
  commercialScore
  score
  grade - Use dataQualityGrade instead
  total

  /api/v1/us/current.json

  posNeg
  total
}
\examples{
tests_daily <- R.COVID.19::covtrck_states_testing()
head(tests_daily)

}
