% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_recovered_daily.R
\name{global_recovered_daily}
\alias{global_recovered_daily}
\title{COVID-19 Global Recovered Counts from John Hopkins}
\usage{
global_recovered_daily()
}
\value{
A datarame/tibble
}
\description{
This functions pulls global recovered COVID-19 data from John Hopkins
  GitHub repository downloads it and transpose it so that it is tidy.  No
  data is stored within the package, so the data should continuously update
  over time as long as the links do not change.
}
\examples{
\dontrun{
confirmed <- R.COVID.19::global_recovered_daily()
}

}
