% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_vac_us_by_state.R
\name{owid_vac_us_by_state}
\alias{owid_vac_us_by_state}
\title{Vaccinations for COVID-19 Counts for US States from Our World in Data}
\usage{
owid_vac_us_by_state()
}
\value{
A dataframe/tibble
}
\description{
This functions pulls COVID-19 vaccine data for US states from
  Our World in Data website GitHub repository and downloads it.
  No data is stored within the package, so the data should continuously
  update over time as long as the links do not change.
}
\details{
Website:

    - https://github.com/owid/covid-19-data/tree/master/public/data/vaccinations

    - https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/vaccinations/us_state_vaccinations.csv
}
\examples{
confirmed_st_daily <- R.COVID.19::owid_vac_us_by_state()
head(confirmed_st_daily)

}
