% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_counties_daily.R
\name{us_counties_daily}
\alias{us_counties_daily}
\title{COVID-19 Counts for US Counties from The New York Times}
\usage{
us_counties_daily()
}
\value{
A datarame/tibble
}
\description{
This functions pulls death COVID-19 confirmed and death data
  for US counties from The New York Times GitHub repository downloads it.
  No data is stored within the package, so the data should continuously
  update over time as long as the links do not change.
}
\examples{
\dontrun{
confirmed <- R.COVID.19::us_counties_daily()
}

}
