% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_geo_confirmed_daily.R
\name{us_geo_confirmed_daily}
\alias{us_geo_confirmed_daily}
\title{COVID-19 US Confirmed Counts from John Hopkins}
\usage{
us_geo_confirmed_daily()
}
\value{
A datarame/tibble
}
\description{
This functions pulls US confirmed COVID-19 data from John Hopkins
  GitHub repository downloads it and transpose it so that it is tidy.  No
  data is stored within the package, so the data should continuously update
  over time as long as the links do not change.
}
\examples{
\dontrun{
confirmed <- R.COVID.19::us_geo_confirmed_daily()
}

}
