% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finra_shelter_in_place.R
\name{finra_shelter_in_place}
\alias{finra_shelter_in_place}
\title{FINRA Shelter in Place Dates}
\usage{
finra_shelter_in_place()
}
\value{
A dataframe/tibble
}
\description{
This functions pulls dates for shelter in place order by states with notes
  and start date and end dates, as well as, extention dates from the FINRA
  COVID-19 website. No Parameters
}
\examples{
\dontrun{
stay_orders <- R.COVID.19::finra_shelter_in_place()
}

}
